should work but i need some break for now...        TODO: make this at least blink without the button
rector:(macros) The size calculation should be asking for a wave.
This is known behavior if you need to make the waveshape change on the
fly. Try to make a nice wave creation function and put the data in the
size calculation macro for modularity later.

lets try this on the scope and see what will give.
could be better if the size was in a variable in settings struct or
something. We'll see.

fut:(invest) Should be checked with scope.
I need a scope... you need a scope... who has a scope??

we need the gpio stuff to start the exti's.

rector:(:memory/reduction) Try to have better mem.
This is due to the fact that i will need to have bigger data lengths
while maintain the resolution in higher frequencies. If i give a struct
each time i want to reinit the TIM6 to change output frequency or
something. This needs to be solidified and test different frequency
values for the ARR and check the corespondance in relation to the
length. The bigger issue is i need to reliably measure these things but
no tools are in possession. @zasbe tell me that you can do something.. ;p

feat:(:buttons) Semi working only the wait func.
This is some wierdness with the callback function and should be checked.
all the timer things are not checked and dma is independent from the
loops.

rector:(dac) Create a concise way to handle dac.
Adding a struct to represent the dac settings and specific attributes
with the intend to be more clear on where something should be put.
All these are in vain if i don't make the selection possible... should
check the runtime capabilities... I cant do that in the compilation
process with a default signal filled with 0's. This is unprofessional
and unicorn.

feat:(rector) make waves runtime choise.
In order to select the waves @runtime we have to seperate the settings
from the waves things and set a default value for an empty wave for
starters. After change all the config functions to also be reinitialized
to change the waves... I have an idea but the button struct and all the
changes here go towards this.

basic cleanup and removing stale args from functions.
also pin point the issue with =DAC= reinit.

add(:wave/picker) A way to change the DMA request.
This is not complete and not working at the moment. Needs the ability to
check if the data is in the needed format but above all needs to change
the f*ck**g wave already...

fix:(:button) User response is done with isr.
Should test the dac/dma integration with this.