fixup! feat:(stability) Dither removal. We remove the dither as is no longer used by the generator. The reason is that now we stabilize the signal through the interpolation of the data after the acumulator has mapped the index to the OG table of the signal to be served by the DAC. There is also no need for debouncing the potentiometer of the pitch because we want all the values and the range is managed by the mapping function.
Reapply "fixup! feat:(stability) Dither removal. We remove the dither as is no longer used by the generator. The reason is that now we stabilize the signal through the interpolation of the data after the acumulator has mapped the index to the OG table of the signal to be served by the DAC. There is also no need for debouncing the potentiometer of the pitch because we want all the values and the range is managed by the mapping function."

This reverts commit 32e4f83ac2f64994da5c9786c699fd477b0feb20.