fix:(mind fuck naming) This is the current values not from the past.
add:(tesing) We should be clear on how the updates should happen on the DAC.
We may need to be more aggrecive and abord computaiton of the signal all
along but for now this is sufficient.

The rattling is still here but voltage regulation is the rescue!!

fix:(gcc wining)

fix:(bouncing issues) We need this binary relationship to function.
The main reason is there should not be a point in time where those
values are equal.

Then the hole logic brakes. iT is safer to change the one in relation to
the other and leave some of the fuckeries when interrupts are kick in...