feat: :module cleanup and create explicit functionality for gpio driver. Most of the work here is to make correct assumptions for the world.
Until now the gpios where false initializing to the maximum number of
pins. If those where the CV analog pins (which now exceed the digital
gpios) all the arrays smaller than "sizeof(adc_pins)" would bounce of
the edge and may write to other memory when enabling optimizations.

For performance reasons the wave choice is moved from button presses to
encoder clicks. This is to ensure normal operations and a less trouble
due to button presses.