feat: adc Add new channel for second distortion amount point. Use second point to do another form of modulation to the frequency by selecting a second point of the LUT to change the slope L=y/x.
This is the commit message #2:

fix:(ui choice settings) Use a more readable branching point.
We should bit_bang_encoder() not only if the distortion is on anymore
but also if we want to change the waveform.

The encoder changes waves in the menu system now so this should be
explicitly chosen by the settings.

This is the commit message #3:

feat:(ignore) cosmetic