fix: decouple increment from encoder struct. Most of the times the encoders have multiple functionality under different conditions so it is better to compute what we want on the fly.
This removes the increment value from the encoder structure and gives
total freedom to what we compute and what we increment. Better way is to
increment each value at position that way it is not a mystery what
happens to end up at the specific increment.

Fixes also the bug in which moving the increment -> then trying to
increment the other mode (i.e from wave to distortion or vise-versa) was
holding the previeus mode's increment and was incrementing this value.
It was what I have coded but not what I want now after the many months
of iterating over the encoder!